+++ src/error.rs
@@ -1,10 +1,10 @@
 // Copyright (c) SimpleStaking and Tezedge Contributors
 // SPDX-License-Identifier: MIT
 
-use crate::mlvalues::tag;
+use crate::mlvalues::{MAX_FIXNUM, MIN_FIXNUM, tag};
 use crate::mlvalues::{is_block, string_val, tag_val, RawOCaml};
 use crate::value::caml_string_length;
-use std::slice;
+use std::{fmt, slice};
 
 /// An OCaml exception value.
 #[derive(Debug)]
@@ -15,6 +15,25 @@ pub struct OCamlException {
 #[derive(Debug)]
 pub enum OCamlError {
     Exception(OCamlException),
+}
+
+#[derive(Debug)]
+pub enum OCamlFixnumConversionError {
+    InputTooBig(i64),
+    InputTooSmall(i64),
+}
+
+impl fmt::Display for OCamlFixnumConversionError {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            OCamlFixnumConversionError::InputTooBig(n) => {
+                write!(f, "Input value doesn't fit in OCaml fixnum n={} > MAX_FIXNUM={}", n, MAX_FIXNUM)
+            }
+            OCamlFixnumConversionError::InputTooSmall(n) => {
+                write!(f, "Input value doesn't fit in OCaml fixnum n={} < MIN_FIXNUM={}", n, MIN_FIXNUM)
+            }
+        }
+    }
 }
 
 impl OCamlException {